Remove fp32 deps from videoflow

+++ ptlflow/models/videoflow/videoflow_bof.py
@@ -94,8 +94,8 @@ class VideoFlowBOF(BaseModel):
     def initialize_flow(self, img):
         """Flow is represented as difference between two coordinate grids flow = coords1 - coords0"""
         N, C, H, W = img.shape
-        coords0 = coords_grid(N, H // 8, W // 8).to(img.device)
-        coords1 = coords_grid(N, H // 8, W // 8).to(img.device)
+        coords0 = coords_grid(N, H // 8, W // 8, dtype=img.dtype, device=img.device)
+        coords1 = coords_grid(N, H // 8, W // 8, dtype=img.dtype, device=img.device)
 
         # optical flow computed as difference: flow = coords1 - coords0
         return coords0, coords1
@@ -135,7 +135,6 @@ class VideoFlowBOF(BaseModel):
         fmaps = self.fnet(images.reshape(B * N, 3, H, W)).reshape(
             B, N, -1, H // 8, W // 8
         )
-        fmaps = fmaps.float()
         fmap1 = fmaps[:, 0, ...]
         fmap2 = fmaps[:, 1, ...]
         fmap3 = fmaps[:, 2, ...]
